requirements.txt
passlib~=1.7.4
requests~=2.31.0
speedtest-cli~=2.1.3
mysql-connector-python~=8.3.0
pyinstaller~=6.6.0
python-dotenv~=1.0.1